requirements.txt
aiofiles==24.1.0
aiosmtplib==4.0.2
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
playwright==1.55.0
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyPDF2==3.0.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.9.18
python-docx==1.1.2
docxtpl>=0.20.1  # DOCX template rendering for contract generation
docx2pdf>=0.1.8  # DOCX to PDF conversion (requires LibreOffice or MS Word on Windows)
weasyprint>=66.0  # HTML to PDF conversion for contract generation
openai==1.59.5
google-genai>=0.3.0  # For File API (PDF/DOC extraction)
google-generativeai>=0.3.0  # For Text Generation API
reportlab>=4.0.0  # PDF generation for analysis downloads

# LangChain and AI/ML Libraries
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-core>=0.1.0  # Core LangChain components
langchain-text-splitters>=0.0.1  # Text splitting utilities
chromadb>=0.5.3  # Persistent vector store backend for Chroma
# Phase 3 & 4: Document processing and embeddings
pdfplumber==0.11.4
PyMuPDF==1.26.4
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.1.0
# Arabic text processing
arabic-reshaper==3.0.0
python-bidi==0.6.6
numpy>=1.24.0

# Embeddings and Search (Upgraded for Arabic Legal Models)
transformers>=4.35.0  # Required for Arabic BERT models
torch>=2.0.0  # PyTorch for model inference
sentencepiece>=0.1.99  # Required for some tokenizers
faiss-cpu>=1.7.4  # Fast similarity search (use faiss-gpu for GPU support)
sentence-transformers>=2.2.0  # Sentence embeddings (fallback)
# Legal Knowledge Management (Phase 5)
# Note: pgvector is PostgreSQL-specific, using JSON storage for SQLite compatibility
# For future PostgreSQL migration, add: pgvector>=0.2.0
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tiktoken==0.11.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
# uvloop==0.21.0  # Not supported on Windows
watchfiles==1.1.0
websockets>=13.0.0
wheel==0.45.1

# Additional packages found in codebase
easyocr>=1.7.0  # OCR library for Arabic text extraction
opencv-python>=4.8.0  # Computer vision library (cv2)

# Performance optimization dependencies
ijson>=3.2.3  # Incremental JSON parsing for large files
aiofiles>=24.1.0  # Async file operations (already present)